Module amazon_sp_sales::apis::sales_api

Functions

get_order_metrics

Returns aggregated order metrics for given interval, broken down by granularity, for given buyer type. Usage Plan: | Rate (requests per second) | Burst | | –– | –– | | .5 | 15 | The x-amzn-RateLimit-Limit response header returns the usage plan rate limits that were applied to the requested operation, when available. The table above indicates the default rate and burst values for this operation. Selling partners whose business demands require higher throughput may see higher rate and burst values than those shown here. For more information, see Usage Plans and Rate Limits in the Selling Partner API.